fix: Match Response.json() type to JSON.parse()
This patch updates the `Response.json()` type to mirror the result of
Node's `JSON.parse()` interface. Previously, `json()` was defined as
returning `object`; while this may appear correct from a "human"
point-of-view, this is actually the same as saying that the `json()`
always returns an empty Object, which is incorrect.